+++ src/annotation.ts
@@ -1,5 +1,4 @@
 import { CodeLensProvider, CodeLens, CancellationToken, TextDocument, workspace, window, Uri } from 'vscode';
-import { extensionDisabled } from './configuration';
 
 
 const defaultAnnotations = {
@@ -46,9 +45,6 @@ export class AnnotationProvider implements CodeLensProvider {
 	constructor() {}
 
 	public async provideCodeLenses(document: TextDocument, _token: CancellationToken) {
-
-		if (extensionDisabled())
-			return [];
 
 		const annotationsPath = getAnnotationsPath(document.uri.path);
 		const annotations = allAnnotations[annotationsPath];