+++ src/configuration.ts
@@ -1,5 +1,5 @@
 import { workspace } from 'vscode';
 
 const getConfiguration = <T>(key: string) => workspace.getConfiguration('PackageAnnotator').get<T>(key);
-export const extensionDisabled = () => !getConfiguration<boolean>('enabled');
 export const hideWarning = () => getConfiguration<boolean>('hideMissingAnnotationsWarning');
+export const hideButton = () => getConfiguration<boolean>('hideGenerateAnnotationsButton');