+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## v1.0.3 (2/2/2025)
+
+### Added
+- Setting to hide missing annotations warning
+- Setting to hide generate annotations button
+
+### Removed
+- Setting to enable/disable extension
+- Setting to ignore missing annotations warning
+
+### Changed
+- All status bar items will now be hidden when all annotations files exist
+- Updated README
+
+
 ## v1.0.2 (2/2/2025)
 
 ### Changed